AI_Career_Guide/requirements.txt
Flask
flask-cors
pandas
spacy
PyMuPDF
python-docx
scikit-learn
numpy
